Move bulk-loading implementation into a submodule.
Facilitating reverting to the original behavior could be useful, should entity
loading/iteration interfaces change.
